add removeFromHand method
